Fix issues with msgpack-c header inclusion ordering
There seems to be some issues with the order in which msgpack-c
headers are included that can cause strange behavior. In this case
the compiler was unable to resolve the stream operators used for
defining streaming operations for custom enumeration types. I also
fixed all of the msgpack objects used to hold enums as they are no
longer required to have a zone passed to the object constructor.