Generalize a rawsocket server for the wamp router.
To facilitate a wamp router exposing itself as a raw/native endpoint
we need initially added a tcp base server interface. However, we will
also need to be able to support unix domain sockets in the near future.
As a result, the tcp server has been abstracted into a rawsocket server
which supports plugins for different types of listeners. This will make
adding support for rawsocket based connections much simpler.