[dealer] Fix multiple registration bug
We were previously only tracking a single registration id. This meant
that when a session registered multiple procedures and then shutdown
and started up again, we would get error messages upon registration
claiming that the procedure was already registered. Further to this
it was not necessary for procedure registrations to use a weak session
so that has been reverted back to a shared pointer.