[dealer] Fix the invocation lifecycle
Invocations are now cleaned up properly in the event that the
caller or callee session is closed in mid call. Also, if the
callee session closes before it yields an invocation the dealer
will now cancel any outstanding calls associated with the callee
session. This fixes an issue where the call would hang until
timeout if the callee session closed before yielding.