[test] Add MIT copyright headers back into the Python files.
They were removed in commit 5afc6b2 which reorganized and
modified the tests substantially, but not enough analysis has
been done to conclusively determine that no copyright remains
from the original files.

We therefore assume that the original copyright is still in place,
and just add our own in addition to Tavendo's.