Only enable -fdiagnostics-color for GCC on 4.9 or higher.
Turns out the option didn't exist on 4.8 yet.

Ninja needs an "always" setting to show color in the terminal but
automatically removes color codes when piped to a non-terminal.
make works better with an "auto" setting for both cases.
Make it print in color in the most possible circumstances.

Fixes #1.