Add support for native components
This commit adds the necessary infrastructure required to allow
for components to connect natively to bonefish and join a session
with a wamp router. This should allow for some very high performance
messaging as native support was added with a zero copy philosophy. A
message that is sent to bonefish over a native transport is only ever
shallow copied according to the msgpack::object and msgpack::zone
semantics. A native transport implementation is still in progress using
AutobahnCpp and should be too far behind this commit. It will be used to
finish testing. I would consider native support to be experimental at this
stage and it is not yet enabled in the bonefish daemon.

Fixes: #21